fix #11: have a target definition
This introduces an explicit target definition to avoid incompatible
changes. Developers need to open the .target file and hit "Set as active
target platform". The Tycho build uses the target automatically.

Java changes are just to get back to the old style adapter
implementation which requires an explicit cast. Manifest change is to
remove a seemingly unused e4 dependency.

I'm running an installation that uses the binary produced by this change
and it works fine.