upgrade target to Eclipse Juno
To have the e4 optional dependencies available at compilation time, let
us use Juno instead of Indigo as target platform. This does not mean
that eclipse-cs no longer works (or installs) on Indigo. The feature
still has the same (non-versioned) dependencies as before and will
install on any org.eclipse.core >= 3.7.

EMF has been added to the target because the e4.model transitively
requires it. eclipse-cs is independent of it.