releng: have AutoDeriv config for docs folder
When using the AutoDeriv plugin, this automatically set the "docs"
folder as derived, thereby excluding it from file search etc.
Unfortunately the "derived" flag is not part of the workspace itself
(therefore it cannot be versioned).

Developers not using the AutoDeriv plugin are not affected at all by
this change.