advertise additional checks
Have a reference to the update site of the additional checks in our
update site. That way it is automatically added to the list of known
update sites in the local P2 update manager. However, it is initially
disabled, and must be enabled manually by the user (so no additional
checks get installed by this change).

Also change the update site URL inside the update site description
(there is actually no field in the form based category editor showing
this property).