fix #8: modifier order quickfix not working
The offset reposition calculation didn't include the indentation of the
Javadoc, there the offset was wrong by (indentation)*(javadoc lines)
bytes.

To avoid knowing the indentation, the quickfix now calculates the
modifier positions of the AST node instead, since the marker position
must be between the first and last modifier node position.