Update to checkstyle 8.9
This includes the 8.9 release and sources of checkstyle. Version numbers
and documentation of EclipseCS have not been modified yet.

The source file name has changed due to the new release hosting on
github. The classpath and the wiki description have both been adapted to
this new naming scheme (checkstyle-checkstyle-x.y.zip).