update Maven plugins
Update all the Maven plugins which have a fixed version. To simplify the
check for Maven updates in the future, a new shared launch config for
Eclipse has been added.

Still, some standard Maven plugins (and Maven itself) are not versioned,
which can lead to different build results, and should be changed.