fix test classpath
Two of the bundles contain nested tests, which are mixed with the main
sources, since both are on the main classpath. Mark those test sources
as containing tests only.

With this change the tests are shown as a separate test source folder
and can be run directly from eclipse. It is not possible to accidentally
mix test and main sources, because Eclipse has a main/test classpath
separation.

Also make all projects use Linux line terminators for new files by
default, since that seems to be what the majority of files is using
currently.