ci: replace tycho-source-feature-plugin
The previously separate tycho-source-feature-plugin functionality is
included in tycho-source-plugin meanwhile:
https://github.com/eclipse-tycho/tycho/blob/master/RELEASE_NOTES.md#tycho-source-feature-generation-moved-from-tycho-extras-to-tycho-core.
Replacing it removes one warning from the Maven build.

Also move the configuration into the parent POM, since it is independent
of the specific feature being built (i.e. could be reused if another
feature would be built).

And remove some commented out configuration of the p2-metadata plugin.
We wouldn't want to use it with those configuration values anyway.

Verified the source feature in the update site is identical to what it
was before.