ci: upgrade Tycho to 4.0.5
That is the most recent version of Tycho. It requires Java 17 for
running, but can compile Java 11 code.

One of the plugins has been moved from tycho.extras to tycho core,
therefore the groupid is different. Some new tycho plugins have been
added. Even though we don't use them, they need to be configured due to
our strict enforcer checks.

There are new warnings at the very end of the build like `No digest
algorithm is available to verify download`. That's an open issue in
Tycho but it doesn't break the build, see
https://github.com/eclipse-tycho/tycho/issues/1482

Finally the target platform also needs to be updated. Tycho 3.0 resolves
the JUnit classpath container of the .classpath files against the bundle
names introduced in Eclipse 2022-06. Most thirdparty libraries have been
changed from repackaging in Eclipse Orbit to "consume as is" from Maven.
This does not affect the EclipseCS plugin created by the build, since
the new JUnit bundles are only needed at IDE or build time.

The jgit exception at the end of a Maven build can be ignored. It's
fixed already upstream, but will probably only be available for use in
Tycho with the Eclipse 2024-03 release.