infra: convert UI libraries to maven
Convert also the external libraries used by the UI bundle to plain Maven
coordinates. Import the packages provided by the libraries with the
minimum package version matching the current library version.

The version number 1.0 for jfreechart-swt is not an error. The previous
jfreechart-1.0.19-swt was part of the jfreechart-1.0.19 (without SWT)
release. It has been split out into a separate project now with version
number 1.0 (so that new 1.0 is more recent than the old 1.0.19). That's
also where the change of the "experimental" package name comes from.

In a future change we should also upgrade the libraries, but I want to
do that separately, because I remember these libraries have trouble with
the module system and its restrictions.

And finally, the target platform has been split once more as suggested
in previous comments to have the root target file only include other
targets but not reference any library by itself.