Infra: Avoid repeated target file resolution
Due to a bug in Tycho the path towards target platform files is not
normalized. Therefore a/../target and b/../target are considered
different files, even though they are in fact the same.

As a workaround use a path relative to the directory where the maven
build was started in, which is identical for all maven modules. This is
typically not recommended, because it would not work when building a
single module only, but that doesn't work anyway in this project (due to
the checkstyle maven config properties resolution).