Infra: re-enable tycho jgit timestamp provider
The Tycho JGit timestamp provider calculates the timestamp based on the
last change of any file in the module. It's therefore okay to be
different for different child modules, so it was wrong to disable it in
the first place.

The benefit of this timestamp provider is that it leads to the exact
version number each time we re-build the same commit (since that commit
defines the timestamp). Without this, rebuilding the same commit leads
to the _current_ timestamp (build time) being used, leading to different
file names for the same commit.