Issue #492: abort opening more module dialogs after cancel
When adding multiple modules to a configuration, the module
configuration dialog is opened for each of them. If the user did that on
accident, she had to go through as many dialogs as modules were
selected.

With this change after the first cancellation of any dialog, no more
dialogs are opened and no more rules are added.