Infra: make documentation timestamp reproducible
Instead of adding the current time into the documentation sitemap, use
the last change in any file of the docs project. Also use only the date
part of the timestamp, not the full second resolution timestamp (using
only the date is a valid W3C date format, which is used in sitemaps).

That way rebuilding with maven from an existing git commit will no
longer lead to unwanted changes.