Fix icons not being transparent
The checked icons previously didn't have correct transparency. Also they
were created as a single icon with all image content statically composed
together in a single file.

Eclipse supports overlayed icons. Therefore make the checked icons a
composition of the base icon and an overlay for the checkmark. That way
transparency is fixed, both the general icon and overlay adapt to HiDPI
resolution and the checkmark image file can be replaced and updated
independent of the base image file.