infra: downgrade JUnit
Upgrading Maven Surefire to a recent version fails with exceptions in
the process launching the test process. That's because we have 2
versions of JUnit: 5.10 from Maven and 5.7.1 from the Eclipse target
platform. Surefire seems to be confused by that and tries to use the
junit-platform-launcher 5.7 from the target platform to launch the 5.10
JUnit during the Maven run.

To avoid this, downgrade the Maven JUnit version to the latest 5.7.x,
such that both Maven and Eclipse use the same JUnit version.

Also remove some superfluous configuration.