infra: filter duplicated resources
When importing all projects in Eclipse, the Maven modules are imported
as standalone projects, but also appear as child folders of the root
module. That in turn leads to always having each file appear twice when
opening the resource selection dialog, running a file search etc.

Mark the child folders in the root module project as derived, which will
hide them in most file related operations afterwards.