Fix build order
The unpacking previously happened in the prepage-package lifecycle phase
of maven. That's far too late, since other compilation steps require
those sources. I had multiple failing local builds because of this.

Until now this only worked correctly because either
* the eclipse builder would compile the depending sources again when the
maven build unpacked everything
* the maven build was not fast enough in parallel processing
* the sources were still available locally from the last download, when
not using clean in between

Also ensure the order of download and unpack is correct by converting
the download to an Ant task and moving it into the same execution. As a
side effect, this avoids the download progress messages of the previous
download-maven-plugin.