cleanup bundle relations
* Remove additional.bundles from build.properties (except for testing
dependencies). Those are NOT meant for hiding runtime dependencies, but
just development time and testing dependencies.
* Make e4 dependency non-optional. Our minimum target platform always
contains e4.
* Make team provider dependency non-optional. The way the Eclipse CS
code is written, those are required (we would have to defer class
loading if this were an actual optional dependency).
* Add org.eclipse.help dependency to doc bundle. That bundle provides
the help extension and therefore also should depend on the help system.
* remove build property which is only relevant for PDE export. We build
with Tycho.