remove source repackaging
Previously the sources of the checkstyle library were extracted during
the build of the library plugin project, just to make them available in
the source feature on the update site.
This is not necessary at all for the plugin to compile and work, and it
doesn't help anyone who wants to use the sources of our plugin.
Therefore this should just be removed.