Correctly look up metadata from checkstyle marker
Rule metadata is registered only under the short name of the rule. But
markers created for checkstyle violations use the fully qualified name
of the checker class instead. Register the fully qualified name as
alternative name in the metadata, and the lookup from a stored marker
works as expected.

Fixes #397