Add command for running checkstyle
This adds a command for the already existing action to run checkstyle on
the selected resources. That way users can invoke the command also via
Ctrl-3 quick assist or via a key binding.

There is no default key binding, as the normal workflow for users is to
rely on the Eclipse builder and to not invoke this command manually.

fixes #99